+++ src/Opulence/Cryptography/Hashing/IHasher.php
@@ -23,21 +23,19 @@ interface IHasher
      * Gets the hash of a value, which is suitable for storage
      *
      * @param string $unhashedValue The unhashed value to hash
-     * @param array $options The list of algorithm-dependent options
      * @param string $pepper The optional pepper to append prior to hashing the value
      * @return string The hashed value
      * @throws RuntimeException Thrown if the hashing failed
      */
-    public function hash(string $unhashedValue, array $options = [], string $pepper = ''): string;
+    public function hash(string $unhashedValue, string $pepper = ''): string;
 
     /**
      * Checks if a hashed value was hashed with the input options
      *
      * @param string $hashedValue The hashed value to check
-     * @param array $options The list of algorithm-specific options
      * @return bool True if the hash needs to be rehashed, otherwise false
      */
-    public function needsRehash(string $hashedValue, array $options = []): bool;
+    public function needsRehash(string $hashedValue): bool;
 
     /**
      * Verifies that an unhashed value matches the hashed value